Wanna get this shit over with

+++ Examples/Example3/Main.cpp
@@ -58,7 +58,7 @@ class Program {
                 std::swap(bytes[i], bytes[i + 2]);
                 bytes[i + 3] = 255; // Bmp broken???
             }
-            int dim = sqrt((bytes.size() - headerOffset) / 4);
+            int dim = (int)sqrt((bytes.size() - headerOffset) / 4);
             return Texture::load(*graphics, dim, dim, bytes.data() + headerOffset, Texture::Format::RGBA32);
         }
 
@@ -96,7 +96,7 @@ class Program {
                 }
                 bytes[i + 3] = 255;
             }
-            int dim = sqrt((bytes.size() - headerOffset) / 4);
+            int dim = (int)sqrt((bytes.size() - headerOffset) / 4);
             tex = Texture::load(*graphics, dim, dim, bytes.data() + headerOffset, Texture::Format::RGBA32);
             
             vertices = StructuredData(shader2->getVertexLayout(), 4);
